Created new directory for this work, renamed old directory
