+++ tinker/faculty_bio/__init__.py
@@ -1,1 +1,2 @@
-__author__ = 'ces55739'
+
+
