+++ tinker/faculty_bio_old/templates/faculty-bio-confirm-edit.html
@@ -1,0 +1,20 @@
+{% extends "tinker-base.html" %}
+
+{% block title %}Congrats!{% endblock %}
+
+
+
+{% block main_content %}
+
+  <div class="row">
+    <div class="large-12 columns">
+        <h1>Congrats!</h1>
+        <p>You've successfully edited your bio. Your edits have been sent for approval but will be ready to view in 2-3 business days. Thanks for keeping your bio up to date!</p>
+        <p>If you have any questions, you can contact Lauren Pareigat (CAS), Annette Abel (CAPS/GS), or Jenny Vang (Seminary).</p>
+
+
+
+    </div>
+  </div>
+
+{% endblock %}
