+++ tinker/faculty_bio_old/templates/faculty-bio-confirm-new.html
@@ -1,0 +1,20 @@
+{% extends "tinker-base.html" %}
+
+{% block title %}Congrats!{% endblock %}
+
+
+
+{% block main_content %}
+
+  <div class="row">
+    <div class="large-12 columns">
+        <h1>Congrats!</h1>
+        <p>You've successfully created a new bio. Your brand new bio has been sent for approval but will be ready to view in 2-3 business days.</p>
+
+        <p>Thanks for creating your bio!</p>
+
+
+    </div>
+  </div>
+
+{% endblock %}
