+++ tinker/faculty_bio_old/templates/faculty-bio-delete-confirm.html
@@ -1,0 +1,19 @@
+{% extends "tinker-base.html" %}
+
+{% block title %}Success{% endblock %}
+
+{% block page_title %}Bethel University Tinker{% endblock %}
+
+{% block main_content %}
+
+  <div class="row">
+    <div class="large-12 columns">
+
+        <p>
+          Your faculty bio has been deleted. It will be removed from your <a href="https://tinker.bethel.edu">Tinker homepage</a> in a few minutes.
+        </p>
+
+    </div>
+  </div>
+
+{% endblock %}
