+++ tinker/faculty_bio_old/templates/faculty-bio-home.html
@@ -1,0 +1,82 @@
+{% extends "tinker-base.html" %}
+
+{% block title %}Home{% endblock %}
+
+{% block styles %}
+  <style type="text/css">
+  </style>
+{% endblock %}
+
+{% block page_title %}Bethel University Tinker{% endblock %}
+
+{% block main_content %}
+<div class="row">
+  <div class="large-12 columns">
+
+    <p>Below is a list of faculty bios you have access to edit. If you don't see your faculty bio, please contact your web author.
+    </p>
+  </div>
+ </div>
+<hr/>
+<div class="row">
+
+  <div class="large-12 columns">
+    {% if show_create %}
+      <a href="/faculty-bio/edit/new#page-title" class="small button success">Create New Bio</a>
+    {% endif %}
+   {% for form in forms%}
+      <div class="row">
+        <div class="large-6 columns">
+          <p>
+            {% if form.path != '' %}
+              <a href={{ form.path }}>{{ form.title }}</a>
+            {% else %}
+              {{ form.title }}
+            {% endif %}
+            <br/>
+            {% if form.path != '' %}<small><a href={{ form.path }}>{{ form.path }}</a></small>{% endif %}
+          </p>
+        </div>
+        <div class="large-6 columns small-3">
+          <ul class="button-group">
+            {% if form.path != '' %}
+              <li><a href="{{ form.path }}" class="small button success">View</a></li>
+              <li><a href="/faculty-bio/edit/{{ form.id }}#page-title" class="small button">Edit</a></li>
+            {% else %}
+              <li><a class="small button success">Not Approved</a></li>
+            {% endif %}
+            <li><a href="#" data-reveal-id="confirmModal" id="{{ form.id }}" data-faculty-bio-id="{{ form.id }}" class="small button alert delete-press">Delete</a></li>
+           </ul>
+        </div>
+        <hr/>
+      </div>
+    {% endfor %}
+
+
+  </div>
+</div>
+
+
+{#Delete confirm modal#}
+<div id="confirmModal" class="reveal-modal" data-reveal>
+  <h2>Are you sure you want to delete?</h2>
+  <p class="lead">Your faculty bio page will be removed.</p>
+  <a class="modal-confirm" id="modal-confirm" href="/faculty-bio/delete/formid">Delete</a>
+  <a class="close-reveal-modal">&#215;</a>
+</div>
+
+{% endblock %}
+
+{% block scripts %}
+
+{#  Script to update the modal submit to delete the correct form#}
+  <script type="text/javascript">
+    $(".delete-press").click(function(){
+      $("#modal-confirm").attr("href", "/faculty-bio/delete/" + $(this).data('faculty-bio-id'));
+    });
+  $(document).foundation();
+  </script>
+
+{% endblock %}
+
+
