+++ tinker/faculty_bio_old/templates/faculty-bio-in-workflow.html
@@ -1,0 +1,22 @@
+{% extends "tinker-base.html" %}
+
+{% block title %}Pending Approval{% endblock %}
+
+{% block page_title %}Bethel University Tinker{% endblock %}
+
+{% block main_content %}
+
+  <div class="row">
+    <div class="large-12 columns">
+
+        <h3>
+          Edits Pending Approval
+        </h3>
+
+        <p>You recently made edits to your bio and are currently pending approval. Please wait until the changes have been approved before you make additional edits. Go back to your <a href="/faculty-bio">faculty bios</a>.</p>
+
+
+    </div>
+  </div>
+
+{% endblock %}
